+++ NEWS.md
@@ -1,2 +1,6 @@
+
+# 2018-10-13
+* add to practice 12b sample
+
 # 2018-09-29
 * convert @pyimport to pyimport() (06b/07b)